refactoring

+++ e-shop/backend/src/main/java/eshop/backend/request/PriceRequest.java
@@ -2,9 +2,11 @@ package eshop.backend.request;
 
 import lombok.Data;
 
+import java.math.BigDecimal;
+
 @Data
 public class PriceRequest {
     private Long id;
-    private double price;
+    private BigDecimal price;
     private Long VariantId;
 }
